requirements.txt
pytesseract
opencv-python
pandasin